+++ mandelbrot/mandelbrot2-prune.c
@@ -39,6 +39,7 @@
 #endif
 
 #if __STDC_VERSION__ >= 202311L
+#include <stddef.h>
 #define UNREACHABLE unreachable()
 #elif defined(_MSC_VER)
 #define UNREACHABLE __assume(0)
@@ -290,7 +291,7 @@ int main(int argc, char *argv[]) {
   
   img_size &= ~7ul;
   batch_4 = (img_size & 31) == 0;
-  inv = 2. / img_size;
+  inv = 2. / (double)img_size;
   row_size = img_size / 8;
   eq_parts = img_size / N_THREADS;
   remaining = img_size % N_THREADS;
@@ -327,8 +328,8 @@ int main(int argc, char *argv[]) {
     return EXIT_FAILURE;
   }
   for (i = 0; i < img_size; ++i) {
-    reals[i] = inv * i - 1.5;
-    imags[i] = inv * i - 1.;
+    reals[i] = inv * (double)i - 1.5;
+    imags[i] = inv * (double)i - 1.;
   }
   if (batch_4) {
     batch = WORD;
