+++ mandelbrot/mandelbrot3.c
@@ -22,6 +22,7 @@
 #include <limits.h>
 #include <stdlib.h>
 #include <stdio.h>
+#include <stdint.h>
 
 #define N_THREADS 4 /* Adjust to number of hardware threads */
 #if N_THREADS < 1
@@ -42,6 +43,7 @@
 #endif
 
 #if __STDC_VERSION__ >= 202311L
+#include <stddef.h>
 #define UNREACHABLE unreachable()
 #elif defined(_MSC_VER)
 #define UNREACHABLE __assume(0)
@@ -138,7 +140,10 @@ static inline unsigned long mand_long(
   for (i = 0; i < sizeof(unsigned long); ++i) {
     result =
       (result >> CHAR_BIT) |
-      (mand_char(init_r + offset, init_i, masks) << HIGH_SHIFT);
+      (
+       (unsigned long)mand_char(init_r + offset, init_i, masks)
+       << HIGH_SHIFT
+      );
     offset += CHAR_BIT;
   }
   return result;
@@ -281,7 +286,7 @@ int main(int argc, char *argv[]) {
   
   batch_long = sizeof(unsigned char) != sizeof(unsigned long) &&
     img_size % ULONG_BITS == 0;
-  inv = 2. / img_size;
+  inv = 2. / (double)img_size;
   row_size = img_size / CHAR_BIT;
 #ifndef NO_THREADS
   eq_parts = img_size / N_THREADS;
@@ -320,8 +325,8 @@ int main(int argc, char *argv[]) {
   }
 
   for (i = 0; i < img_size; ++i) {
-    reals[i] = inv * i - 1.5;
-    imags[i] = inv * i - 1.;
+    reals[i] = inv * (double)i - 1.5;
+    imags[i] = inv * (double)i - 1.;
   }
 
   if (batch_long) {
